requirements.txt
asgiref==3.3.4
DateTime==4.3
Django==3.2.4
django-crispy-forms==1.11.2
django-geojson==3.2.0
django-leaflet==0.28.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
geojson==2.5.0
gunicorn==20.1.0
numpy==1.20.3
pandas==1.2.4
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
zope.interface==5.4.0
